msgs_good/msgs_bad are now of type uint; _COUNT macros in bogotune unsigned.
git-svn-id: svn://svn.code.sf.net/p/bogofilter/code/trunk/bogofilter@5002 046ba1d7-8659-4f4f-8137-08d775d845e8

+++ src/globals.h
@@ -53,8 +53,8 @@ extern	const char *const system_config_file;
 
 /* for msgcounts.c */
 #define	MSG_COUNT	".MSG_COUNT"
-extern	double		 msgs_good;
-extern	double		 msgs_bad;
+extern	uint		 msgs_good;
+extern	uint		 msgs_bad;
 
 /* for passthrough.c */
 extern size_t msg_register_size;
